Показания счётчиков, начальное показание и потребление выводятся целыми числами и в экранных таблицах отчётов, и в XLSX-выгрузках, включая сводный показатель «Потребление».

+++ app/Reports/ReportSummaryService.php
@@ -602,7 +602,7 @@ class ReportSummaryService
             ],
             'consumption' => [
                 ['key' => 'readings_count', 'label' => 'Показаний', 'type' => 'integer'],
-                ['key' => 'consumption', 'label' => 'Потребление', 'type' => 'decimal'],
+                ['key' => 'consumption', 'label' => 'Потребление', 'type' => 'integer'],
             ],
             default => throw new InvalidArgumentException("Unsupported summary report [{$reportSlug}]."),
         };
